Normalise each entity at the end of the source pipelines.
This is mainly to avoid later problems with signature creation due
to MDA-144. It was previously performed at the end of the render
pipeline (just before signing) but that impacts the responsiveness
of the system as well as increasing peak memory usage. Moving it
back into the source pipeline (and doing it per-entity) hides the
computation from clients as well as reducing query times.